Moved api files in ./controller to ./api - Also added proof-of-concept router structure
