+++ api/users.js
@@ -9,20 +9,32 @@ client.connect();
 usersRouter.get("/", async (req, res) => {
   let error = "N/A";
   let _ret = [];
-  try {
+  try
+  {
     console.log(url);
     const db = client.db("LargeProject");
     const result = await db.collection("Users").find().toArray();
     console.log(result);
-    for (let i = 0; i < result.length; i++) {
+
+    for (let i = 0; i < result.length; i++)
+    {
       _ret.push(result[i]);
     }
   }
   catch(e) {
     error = e.toString();
   }
+
   let ret = { result: _ret, error: error };
   res.status(200).json(ret);
 });
 
+usersRouter.post("/")
+
 module.exports = usersRouter;
+
+const addUserRouter = require("express").Router();
+
+// ...
+
+module.exports = addUserRouter;
